+++ assets/components/widgets/results/results.jsx
@@ -16,13 +16,14 @@ UI.widgets.Results = React.createClass({
 				return setTimeout(checkRefs, 100);
 			}
 
+			var scrollDelay = UI.state.controllerOptions.options.qualifyingResultsDisplayTime.value > 0 ? (UI.state.controllerOptions.options.qualifyingResultsDisplayTime.value/100) * 60 : 10;
 			var diff = self.refs['entries-outer'].clientHeight - self.refs['entries-inner'].clientHeight;
 			setTimeout(function() {
 				if (!self.refs['entries-inner']) {
 					return;
 				}
 				self.refs['entries-inner'].style.top = diff+'px';
-			}, 10 * 1000);
+			}, scrollDelay * 1000);
 		})();
 	},
 	componentWillUnmount: function() {