+++ changelog.md
@@ -1,5 +1,6 @@
 ### Version 1.7.0
 * [Feature] Added 'Code 60' indicator in race session as used in the 24H series.
+* [Feature] Small UI improvements to the beta broadcaster UI.
 
 ### Version 1.6.1
 * [Bug Fix] Prevented an error being thrown if no local description exists for a track.