+++ changelog.md
@@ -1,6 +1,7 @@
 ### Version 1.7.0
 * [Feature] Added 'Code 60' indicator in race session as used in the 24H series.
 * [Feature] Small UI improvements to the beta broadcaster UI.
+* [Feature] Set qualifying results table scroll delay to 60% of the qualifying table display length instead of a constant seconds.
 
 ### Version 1.6.1
 * [Bug Fix] Prevented an error being thrown if no local description exists for a track.