+++ changelog.md
@@ -8,6 +8,8 @@
 * [Feature] Added damage indicator to beta broadcast panel UI.
 * [Feature] Added mandatory pitstop status to beta broadcast panel UI.
 * [Feature] Condensed camera controls in the beta broadcast panel UI.
+* [Feature] Added tyre icons to the beta broadcast panel UI with green icon for series using dedicated tyres.
+* [Tooling] Switched to new user info endpoint with smaller payloads to improve rendering of user info on screen.
 
 ### Version 1.6.1
 * [Bug Fix] Prevented an error being thrown if no local description exists for a track.