+++ changelog.md
@@ -18,6 +18,7 @@
 * [Feature] Added commentary names widget, can be set in the control option UI.
 * [Dev] Changed window reloading key from 'r' to 'F5' to facilitate typing of commentary names.
 * [Feature] Reduced default resolution scaling by 1px.
+* [Feature] Added PTP remaining info to driver widget for classes with PTP.
 
 ### Version 1.6.1
 * [Bug Fix] Prevented an error being thrown if no local description exists for a track.